feat: move some code
